requirements.txt
ccxt
Pillow
#selenium
bs4
